pyproject.toml
[tool.poetry]
name = "dbj7"
version = "0.1.0"
description = ""
authors = ["timelessnesses <mooping3roblox@gmail.com>", "dragmine149 <>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.0.1"
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
Flask = "^2.1.2"
orjson = "^3.7.3"
python-dateutil = "^2.8.2"
parsedatetime = "^2.6"
jishaku = "^2.5.0"
psutil = "^5.9.1"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
autoflake = "^1.4"
flake8 = "^4.0.1"
watchdog = "^2.1.9"
pylint = "^2.14.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
